transfer permutations to gameObject
